Raise errors on deprecation warnings
Ruby 2.7 introduce a couple of deprecation warnings around [positional and keyword arguments](https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/), in preparation for Ruby 3. So far we only showed the deprecation warning (when running tests for example). Not anymore, because we missed some of them already. We now raise an exception in that case, so you have to fix them before they sneak into investapp.